+++ lib/src/services/yust_file_service_flutter.dart
@@ -20,6 +20,10 @@ class YustFileService {
     }
   }
 
+  YustFileService.mocked() : _fireStorage = FirebaseStorage.instance {
+    throw UnsupportedError('Not supported in Flutter Environment');
+  }
+
   final FirebaseStorage _fireStorage;
 
   Future<String> uploadFile(
@@ -98,5 +102,39 @@ class YustFileService {
     for (final file in fileList.items) {
       await file.delete();
     }
+  }
+
+  Future<List<Object>> getFilesInFolder({required String path}) async {
+    throw YustException('Not implemented for flutter');
+  }
+
+  Future<List<Object>> getFileVersionsInFolder({required String path}) async {
+    throw YustException('Not implemented for flutter');
+  }
+
+  Future<Map<String?, List<Object>>> getFileVersionsGrouped(
+      {required String path}) async {
+    throw YustException('Not implemented for flutter');
+  }
+
+  Future<String?> getLatestFileVersion(
+      {required String path, required String name}) async {
+    throw YustException('Not implemented for flutter');
+  }
+
+  Future<String?> getLatestInvalidFileVersion({
+    required String path,
+    required String name,
+    DateTime? beforeDeletion,
+    DateTime? afterDeletion,
+  }) async {
+    throw YustException('Not implemented for flutter');
+  }
+
+  Future<void> recoverOutdatedFile(
+      {required String path,
+      required String name,
+      required String generation}) async {
+    throw YustException('Not implemented for flutter');
   }
 }