+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 3.23.6 - 2024-07-18
+
+- Add support for PITR in YustDatabaseService
+
+## 3.23.5 - 2024-07-16
+
+- Improvement to YustDatabaseService
+
+## 3.23.4 - 2024-07-11
+
+- Fixes for mocked file service
+
+## 3.23.3 - 2024-07-10
+
+- Add mocked file service
+
 ## 3.23.2 - 2024-07-01
 
 - Fix download URL
@@ -13,7 +29,7 @@
 ## 3.22.1 - 2024-06-07
 
 - Improvements to YustFileService
-  
+
 ## 3.21.2 - 2024-06-07
 
 - Fixes & improvements for YustDatabaseService
@@ -21,7 +37,7 @@
 ## 3.22.0 - 2024-05-30
 
 - Make yust ready for Flutter 3.22.1
-  
+
 ## 3.21.1 - 2024-05-15
 
 - Support DB Statistics in mocked DB Service